Time: 6 ms (74.24%), Space: 23 MB (24.44%) - LeetHub

+++ C++/0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -1,7 +1,7 @@
 class Solution {
 private:
     bool finish(vector<int>& piles, int h, int k){
-        long long hours=0;
+        int hours=0;
         for(auto& p:piles){
             hours+=p/k;
             if(p%k!=0) hours++;
